requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
uvicorn
fastapi[all]
langchain
langchain_groq
langchain_community
langchain-core
langgraph
ipykernel
python-dotenv
celery
redis
beautifulsoup4
requests
sendgrid
imap-tools
pydantic
